refactor: form schema & shared types
