+++ src/domains/tasks/MakeOffer/api/createOffer.tsx
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { MakeAnOfferFormState } from "../hooks";
+import { MakeAnOfferFormState } from "../types";
 
 export const createOffer = async (
   taskId: string,
