+++ src/domains/tasks/MakeOffer/components/MakeOfferForm.tsx
@@ -1,10 +1,9 @@
 import { Button, Form, FormLayout } from "@shopify/polaris";
-import {
-  MakeAnOfferFormState,
-  MakeAnOfferSubmitHandler,
-  useMakeAnOfferForm,
-} from "../hooks/useMakeOfferForm";
+
 import { TextField } from "@ComponentsV2/components/TextField";
+
+import { useMakeAnOfferForm } from "../hooks/useMakeOfferForm";
+import { MakeAnOfferFormState, MakeAnOfferSubmitHandler } from "../types";
 
 type MakeAnOfferFormProps = {
   submitHandler: MakeAnOfferSubmitHandler;
