+++ src/domains/tasks/MakeOffer/hooks/useMakeOfferForm.tsx
@@ -1,20 +1,7 @@
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
-import { z } from "zod";
-
-export const formSchema = z.object({
-  offerAmt: z.coerce.number({ required_error: "Offer Amount is required" }),
-  comments: z
-    .string({ required_error: "Description is required" })
-    .min(25, "Comments must have 25 or more characters")
-    .max(1500, "Comments cannot have more than 1500 characters"),
-});
-
-export type MakeAnOfferFormState = z.infer<typeof formSchema>;
-
-export type MakeAnOfferSubmitHandler = (
-  data: MakeAnOfferFormState
-) => Promise<void>;
+import { MakeAnOfferFormState } from "..";
+import { MakeAnOfferSubmitHandler, formSchema } from "../types";
 
 export const useMakeAnOfferForm = (submitHandler: MakeAnOfferSubmitHandler) => {
   const {
