+++ src/domains/tasks/MakeOffer/index.tsx
@@ -1,2 +1,2 @@
 export { MakeAnOfferModal } from "./components";
-export type { MakeAnOfferFormState } from "./hooks";
+export type { MakeAnOfferFormState } from "./types";
